rebulid day2

+++ modules/notifikasi/read.php
@@ -4,7 +4,7 @@ include '../../includes/koneksi.php';
 
 $sql = "SELECT n.*, a.nama_barang, a.kode_barang 
         FROM notifikasi n
-        JOIN aset_barang a ON n.aset_id = a.id
+        LEFT JOIN aset_barang a ON n.aset_id = a.id
         ORDER BY n.tanggal_notifikasi ASC";
 $result = mysqli_query($conn, $sql);
 ?>
@@ -20,8 +20,9 @@ $result = mysqli_query($conn, $sql);
       <thead>
         <tr>
           <th>No</th>
+          <th>Kode Barang</th>
           <th>Nama Barang</th>
-          <th>Tipe</th>
+          <th>Tipe Notifikasi</th>
           <th>Tanggal</th>
           <th>Status</th>
           <th>Aksi</th>
@@ -31,23 +32,32 @@ $result = mysqli_query($conn, $sql);
         <?php
         $no = 1;
         $today = new DateTime();
-        while ($row = mysqli_fetch_assoc($result)) {
-          $tgl = new DateTime($row['tanggal_notifikasi']);
-          $diff = $today->diff($tgl)->days;
-          $color = ($tgl < $today) ? 'red' : (($diff <= 7) ? 'yellow' : 'green');
 
-          echo "<tr>
-                  <td>{$no}</td>
-                  <td>{$row['nama_barang']}</td>
-                  <td>{$row['tipe_notifikasi']}</td>
-                  <td>{$row['tanggal_notifikasi']}</td>
-                  <td><span class='alert {$color}'>{$row['status']}</span></td>
-                  <td>
-                    <a href='update.php?id={$row['id']}' class='btn'>Edit</a>
-                    <a href='delete.php?id={$row['id']}' class='btn red' onclick='return confirm(\"Hapus notifikasi ini?\")'>Hapus</a>
-                  </td>
-                </tr>";
-          $no++;
+        if (mysqli_num_rows($result) > 0) {
+          while ($row = mysqli_fetch_assoc($result)) {
+            $tgl = new DateTime($row['tanggal_notifikasi']);
+            $diff = $today->diff($tgl)->days;
+            $color = ($tgl < $today) ? 'red' : (($diff <= 7) ? 'yellow' : 'green');
+
+            $nama_barang = $row['nama_barang'] ?? '<i>Data aset dihapus</i>';
+            $kode_barang = $row['id'] ?? '-';
+
+            echo "<tr>
+                    <td>{$no}</td>
+                    <td>{$kode_barang}</td>
+                    <td>{$nama_barang}</td>
+                    <td>{$row['tipe_notifikasi']}</td>
+                    <td>{$row['tanggal_notifikasi']}</td>
+                    <td><span class='alert {$color}'>{$row['status']}</span></td>
+                    <td>
+                      <a href='update.php?id={$row['id']}' class='btn'>Edit</a>
+                      <a href='delete.php?id={$row['id']}' class='btn red' onclick='return confirm(\"Hapus notifikasi ini?\")'>Hapus</a>
+                    </td>
+                  </tr>";
+            $no++;
+          }
+        } else {
+          echo "<tr><td colspan='7' style='text-align:center;'>Belum ada notifikasi.</td></tr>";
         }
         ?>
       </tbody>
